backend/pyproject.toml
[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "C",   # flake8-comprehensions
    "B",   # flake8-bugbear
]
ignore = [
    "E501", # line too long, let formatter handle it
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[dependency-groups]
dev = [
    "fastapi[standard]>=0.100.0",
    "httpx>=0.24.0",
    "numpy>=1.24.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
    "redis>=5.0.0",
    "ruff>=0.1.0",
    "supabase>=2.0.0",
    "uvicorn[standard]>=0.23.0",
    "PyJWT[crypto]>=2.8.0,<3.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = [
  "."
]

[project]
name = "growmate-backend"
version = "0.1.0"
requires-python = ">=3.13"
